Moved tests to fold tests
